# server/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VPS CI/CD Automation Runner API",
    "description": "Webhook-driven git sync, automated zero-downtime deployment runner, and telemetry platform for self-hosted VPS instances.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
...
  "paths": {
...
    "/api/notifications/channels": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List configured notification channels",
        "responses": {
          "200": {
            "description": "List of channels"
          }
        }
      },
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Create notification channel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "provider"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "provider": {
                    "type": "string",
                    "enum": [
                      "slack",
                      "discord",
                      "telegram",
                      "webhook"
                    ]
                  },
                  "webhook_url": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Channel created"
          }
        }
      }
    },
    "/api/notifications/channels/{id}/test": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Dispatch test alert to notification channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test sent"
          }
        }
      }
    },
...
    "/api/hooks/{hookToken}": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Public inbound git webhook receiver",
        "parameters": [
          {
            "name": "hookToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ping or skipped processed"
          },
          "202": {
            "description": "Trigger enqueued"
          },
          "401": {
            "description": "Signature verification failed"
          }
        }
      }
    },
...
    "/api/notifications/channels/{id}": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Update a notification channel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "provider": {
                    "type": "string"
                  },
                  "webhook_url": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object"
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Channel updated"
          },
          "404": {
            "description": "Channel not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Delete a notification channel",
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "404": {
            "description": "Channel not found"
          }
        }
      }
    },
...
  },
...
}